feat(due): Rename the component names from `ons-*` to `v-ons-*` due to the spec of Vue.js 2.
