+++ bindings/vue/src/components/VOnsSwitch.vue
