+++ bindings/vue/src/components/VOnsTabbar.vue
