+++ bindings/vue/src/components/index.js
@@ -1,15 +1,15 @@
-import OnsNavigator from './OnsNavigator.vue';
-import OnsBackButton from './OnsBackButton.vue';
-import OnsTabbar from './OnsTabbar.vue';
-import OnsSwitch from './OnsSwitch.vue';
-import OnsPullHook from './OnsPullHook.vue';
-import OnsSplitterSide from './OnsSplitterSide.vue';
+import VOnsNavigator from './VOnsNavigator.vue';
+import VOnsBackButton from './VOnsBackButton.vue';
+import VOnsTabbar from './VOnsTabbar.vue';
+import VOnsSwitch from './VOnsSwitch.vue';
+import VOnsPullHook from './VOnsPullHook.vue';
+import VOnsSplitterSide from './VOnsSplitterSide.vue';
 
 export {
-  OnsNavigator,
-  OnsBackButton,
-  OnsTabbar,
-  OnsSwitch,
-  OnsPullHook,
-  OnsSplitterSide
+  VOnsNavigator,
+  VOnsBackButton,
+  VOnsTabbar,
+  VOnsSwitch,
+  VOnsPullHook,
+  VOnsSplitterSide
 };
